task 1

+++ main.h
@@ -7,5 +7,5 @@
 void _putchar(char c);
 void print_string(char *str, int *nb_printed);
 int _printf(const char *const format, ...);
-void print_int(int n, int *nb_printed);
+void print_int(long int n, int *nb_printed);
 #endif
